refactor(deco): split Decoration and Surface
Split the Surface from the Decoration to make it compatible with the
wayfire 0.8 API.

Remove the surface of a window if its decoration is meant to be hidden
(either toggled manually or when in fullscreen). It should improve
performances by allowing Wayfire to perform a scan out when possible (on
fullscreen windows, for example).